fixed unreserve, updated reserve

+++ Unreserve.php
@@ -11,13 +11,14 @@
 	include_once ('includes/create-hotbar.php');
 	include_once ('includes/create-footer.php');
 	include_once ('includes/start-session.php');
+	require_once ('includes/config.php');
 ?>
 
 <!DOCTYPE html>
 <html lang="en">
 <head>
 	<meta charset="UTF-8">
-    	<title>About | Sharron Books</title>
+    	<title>Unreserve | Sharron Books</title>
 	<meta name="viewport" content="width=device-width,initial-scale=1">
 	<link rel="stylesheet" href="css/main.css">
 	<title>Unreserve</title>
@@ -35,37 +36,35 @@
 			/* Header bar at the top of the each page containing a series of links and the logo */
 			create_home_header();
 		?>
-	
-	<div class="backgroundfix4">
-		<div class="container">  
-			<div class="main">
-				<h1>Unreserved</h1>
-				<?php
-				echo "<div class='Form2'><h2>Book has been unreserved if code was correct.</h2></div>";
-				?>
-			</div>
-		</div>
-	</div>
-				
-	<?php
-									
-		require('includes/config.php');
-												
-		$Query = $db->Query(sprintf("DELETE FROM bookreserve 
-										WHERE `bookreserve`.`BookID` = '%s'", 
-										$db->escape_string($_POST['BookID'])));
-											
-		echo "<br>";
-		echo "<div class='Form2'><h2>Check your account if the book has been unreserved.</h2></div>";
-		echo "<br>";
 		
-		echo "<div class='Form'><h3><a href='profile.php'>View your account</a> <br></h3></div>";
-		echo "<div class='Form'><h3><a href='includes/logout.php'>Want to log out?</a> <br></h3></div>";
+		<div class="content">
+		<?php
 
-	?>
-	
-	<br><br>
-	
+			// check if book exists
+			$reservationCheckQuery = $db->Query(sprintf("SELECT * FROM bookreserve WHERE BookID='%s' HAVING UserID='%s'", $_POST['BookID'], $_SESSION['login_user']));
+			if ($reservationCheckQuery->num_rows < 1)
+			{
+				// Reservation has already been made
+				echo "<h3> Reservation removal unsuccessful </h3>";
+				echo "<p> The book you selected ( ID number: " . $_POST['BookID'] . " ) Has not been reserved by you. </p>";
+				echo "<p> Click <a href='profile.php'> here </a> to view your active reservations. </p>";
+			}
+			else
+			{
+				// Reservation can be completed
+				$Query = $db->Query(sprintf("DELETE FROM bookreserve WHERE (BookId,UserID)=('%s','%s');", $_POST['BookID'], $_SESSION['login_user']));
+				echo "<h3> Reservation removal successful </h3>";
+				echo "<p> The book you selected ( ID number: " . $_POST['BookID'] . " ) Has  been removed from your reservations. </p>";
+				echo "<p> Click <a href='profile.php'> here </a> to view your active reservations. </p>";
+			}
+
+									
+			
+			
+
+		?>
+		</div>
+		</div> <!-- Lined up -->
 	<?php
 		/* Footer at the end of the page that displays some basic website info */
 		create_footer();
